+++ src/popup/App.css
@@ -1,58 +1,22 @@
-*,
-*::before,
-*::after {
-  box-sizing: border-box;
-}
-
-body {
-  font-family: 'Open Sans', sans-serif;
-}
-
 input,
 button {
   font-family: inherit;
 }
 
 .app {
-  width: 370px;
-  height: 430px;
   background: #eeeeee;
   display: flex;
+  height: 100%;
   flex-direction: column;
   align-items: stretch;
   padding: 32px;
 }
 
-.toolbar {
-  width: 100%;
-  display: flex;
-  margin-bottom: 10px;
-}
-
-.add-btn {
-  background: #5b98f4;
-  border: none;
-  box-shadow: 0px 1px 15px rgba(0, 0, 0, 0.05), 0px 1px 2px rgba(0, 0, 0, 0.1);
-  color: #fff;
-  height: 32px;
-  width: 32px;
-  border-radius: 5px;
-  cursor: pointer;
-}
-
-.tip {
-  text-align: center;
-  padding: 9px 10px;
-  border-radius: 5px;
-}
-
-.tip-warning {
-  background-color: #fbf3c7;
-  color: #734418;
-}
-
-.btn-info {
-  position: fixed;
-  bottom: 5px;
-  right: 5px;
+.small-header {
+  text-transform: uppercase;
+  font-size: 10px;
+  font-weight: 700;
+  letter-spacing: 0.035em;
+  margin-bottom: 8px;
+  color: #707070;
 }
