--- src/popup/Input.css
@@ -1,15 +1,0 @@
-.input:focus {
-  outline: none;
-  border: 2px solid rgb(98, 184, 255);
-  padding: 5px 0 5px 11px;
-}
-
-.input.valid {
-  border: 2px solid rgb(98, 255, 137);
-  padding: 5px 0 5px 11px;
-}
-
-.input.invalid {
-  border: 2px solid rgb(255, 98, 98);
-  padding: 5px 0 5px 11px;
-}
