+++ src/popup/components/table/TableEntry.css
