+++ src/popup/components/tip/Tip.css
@@ -1,0 +1,12 @@
+.tip {
+  display: flex;
+  align-items: center;
+  font-size: 14px;
+  padding: 8px 16px;
+}
+
+.tip-warning {
+  background-color: #f0e3b7;
+  border-radius: 5px;
+  color: #5c4906;
+}
